常规补丁

+++ QuantStudio/FactorDataBase/FactorTools.py
@@ -100,7 +100,7 @@ def fix(f, **kwargs):
     return PointOperation(kwargs.pop("factor_name", str(uuid.uuid1())),Descriptors,{"算子":_fix,"参数":Args,"运算时点":"多时点","运算ID":"多ID"}, **kwargs)
 def _isin(f,idt,iid,x,args):
     Data = _genOperatorData(f,idt,iid,x,args)[0]
-    return np.isin(Data, args["OperatorArg"]["test_elements"]).astype(float)
+    return np.isin(Data, args["OperatorArg"]["test_elements"])
 def isin(f, test_elements, **kwargs):
     Descriptors,Args = _genMultivariateOperatorInfo(f)
     Args["OperatorArg"] = {"test_elements":test_elements}
